requirements.txt
typer>=0.12
pydantic>=2.8
pyyaml>=6.0
python-dotenv>=1.0
rich>=13.7
httpx>=0.27
# 若使用 OpenAI 兼容端点（包括官方/第三方），推荐：
openai>=1.40
google-generativeai>=0.7